First cut of module support in Gru
A module is a configuration written in HCL or JSON which contains
resources, variables, etc. Within a module we can also import
other modules, thus being able to write small and re-usable pieces.

For a module to be discovered in Gru it needs be present in the
module path.

Example of a module in Gru looks like this.

```hcl
import {
  module = [
    "foo",
    "bar",
  ]
}

resource "pacman" {
  name = "tmux"
}
```

The above simple module imports `foo` and `bar` modules and also
declares one resource of type "pacman" for "tmux"