module: initial commit of module.ImportGraph() function
module.ImportGraph() can be used to create a DAG graph of the
module imports for a given module.

The resulting graph can then be topologically sorted in order to
determine the proper ordering of loading and processing modules or
simply to detect whether we have circular imports in our modules.