Refactor task processing
Instead of serializing the whole catalog from the client
and then send it over to minions we now just send the
module name as part of the task.

Since the minions are able to sync module and data files
from an upstream Git repo, they are now able to load a
catalog from the own site's directory and process it.